+++ ng/ngvec.h
@@ -15,6 +15,7 @@ namespace ng {
 	double len (const Vec2& a);
 	double trace (const Mat2& A);
 	double det (const Mat2& A);
+	double det (double a, double b, double c, double d);
 	double dot (const Vec2& a, const Vec2& b);
 	Vec2 unit (const Vec2& a);
 	Vec2 project (const Vec3& a);
@@ -30,20 +31,27 @@ namespace ng {
 	double det (const Mat3& A);
 	double dot (const Vec3& a, const Vec3& b);
 	Vec3 unit (const Vec3& a);
-	Vec3 augment (const Vec2& a);
-	Mat3 augment (const Mat2& A, const Vec2& b);
+	Vec3 affine (const Vec2& a);
+	Mat3 affine (const Mat2& A, const Vec2& b);
 	Mat3 transpose (const Mat3& A);
+	Mat3 minor (const Mat3& A);
 	Mat3 adj (const Mat3& A);
 	Mat3 inv (const Mat3& A);
 	
-	// affine transforms in R2
-	Mat3 translate (const Vec2& p);
-	Mat3 basis (const Vec2& i, const Vec2& j);
+	// affine transforms for R2
+	Mat3 scale (double i, double j);
+	Mat3 scale (double srci, double srcj, double desti, double destj);
+	Mat3 origin (const Vec2& p); // change of origin
+	Mat3 origin (double x, double y);
+	Mat3 basis (const Vec2& i, const Vec2& j); // change of basis
+	Mat3 basis (double i, double j);
 	Mat3 rotate (double a);
-	Mat3 space (const Mat3& translate);
-	Mat3 space (const Mat3& translate, const Mat3& basis);
-	Mat3 space (const Mat3& translate, const Mat3& basis, const Mat3& rotate);
-	Mat3 portal (const Mat3& src_space, const Mat3& dest_space);
+	Mat3 space (const Mat3& origin);
+	Mat3 space (const Mat3& origin, const Mat3& basis);
+	Mat3 space (const Mat3& origin, const Mat3& basis, const Mat3& rotate);
+	Mat3 space (double x, double y);
+	Mat3 space (double x, double y, double i, double j);
+	Mat3 space (double x, double y, double i, double j, double a);
 	
 	// vector in R2
 	class Vec2 {
@@ -58,10 +66,6 @@ namespace ng {
 		Vec2 operator+ (const Vec2& b) const;
 		Vec2 operator* (double x) const;
 		double operator* (const Vec2& b) const; // dot(this,b)
-		double dot (const Vec2& b) const; // dot(this,b)
-		double sqlen () const; // sqlen(this)
-		double len () const; // len(this)
-		Vec2 unit () const; // unit(this)
 	};
 	
 	// matrix in R2x2
@@ -84,24 +88,9 @@ namespace ng {
 		Mat2 operator* (double x) const;
 		Vec2 operator* (const Vec2& b) const;
 		Mat2 operator* (const Mat2& B) const;
-		double trace () const; // trace(this)
-		double det () const; // det(this)
-		Mat2 transpose () const; // transpose(this)
-		Mat2 adj () const; // adj(this)
-		Mat2 inv () const; // inv(this)
 	};
 	
-	// might add this if needed for physics
-	// tensor in R2x2x2
-	/*
-	class Tensor2 {
-	public:
-		Mat2 r;
-		Mat2 s;
-	}
-	*/
-	
-	// vector in R3, or augmented vector in R2
+	// vector in R3, or affine vector for R2
 	class Vec3 {
 	public:
 		double x;
@@ -110,25 +99,27 @@ namespace ng {
 		Vec3 ();
 		Vec3 (double x, double y, double z);
 		Vec3 (const Vec3& a);
-		void set (double x, double y);
+		void set (double x, double y, double z);
 		Vec3 operator+ (double x) const;
 		Vec3 operator+ (const Vec3& b) const;
 		Vec3 operator* (double x) const;
 		double operator* (const Vec3& b) const; // dot(this,b)
-		double dot (const Vec3& b) const; // dot(this,b)
-		double sqlen () const; // sqlen(this)
-		double len () const; // len(this)
-		Vec3 unit () const; // unit(this)
 	};
 	
-	// matrix in R3x3, or affine transformation in R2
+	// matrix in R3x3, or affine matrix for R2x2
 	class Mat3 {
 	public:
-		Vec3 a; Vec3 b; Vec3 c;
+		double a; double b; double c;
+		double d; double e; double f;
+		double g; double h; double i;
 		Mat3 ();
-		Mat3 (const Vec3& a, const Vec3& b, const Vec3& c);
+		Mat3 (double a, double b, double c,
+			double d, double e, double f,
+			double g, double h, double i);
 		Mat3 (const Mat3& A);
-		void set (const Vec3& a, const Vec3& b, const Vec3& c);
+		void set (double a, double b, double c,
+			double d, double e, double f,
+			double g, double h, double i);
 		void set_cols (const Vec3& a, const Vec3& b, const Vec3& c);
 		void set_rows (const Vec3& a, const Vec3& b, const Vec3& c);
 		Vec3 col1 () const;
@@ -142,11 +133,6 @@ namespace ng {
 		Mat3 operator* (double x) const;
 		Vec3 operator* (const Vec3& b) const;
 		Mat3 operator* (const Mat3& B) const;
-		double trace () const; // trace(this)
-		double det () const; // det(this)
-		Mat3 transpose () const; // transpose(this)
-		Mat3 adj () const; // adj(this)
-		Mat3 inv () const; // inv(this)
 	};
 	
 } // ng